50 fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int i;
-	int n1 = 0;
-	int n2 = 1;
-	int n3;
+	long int n1 = 0;
+	long int n2 = 1;
+	long int n3;
 
 	for (i = 1; i <= 50; i++)
 	{
 		n3 =  n2 + n1;
 
-		printf("%d", n3);
+		printf("%lu", n3);
 
 		n1 = n2;
 		n2 = n3;
